663 - Clean src/routers/rootRouter.ts
