+++ src/dispatchers/frontendDeployments.ts
@@ -2,7 +2,6 @@ import {getFrontendDeploymentStatus} from 'api/frontendDeployments';
 import {DEPLOYMENT_TIMESTAMP} from 'constants/localStorage';
 import {setCurrentDeployment, setPollingEnabled, setUpdateAvailable} from 'store/frontendDeployments';
 import {AppDispatch} from 'types';
-import {FrontendDeployment} from 'types/frontendDeployment';
 
 export const checkForDeploymentUpdate = () => async (dispatch: AppDispatch) => {
   try {
@@ -25,17 +24,5 @@ export const checkForDeploymentUpdate = () => async (dispatch: AppDispatch) => {
   } catch (error) {
     dispatch(setPollingEnabled(true));
     throw error;
-  }
-};
-
-export const handleDeploymentUpdate = (deployment: FrontendDeployment) => (dispatch: AppDispatch) => {
-  const storedTimestamp = localStorage.getItem(DEPLOYMENT_TIMESTAMP);
-
-  if (!storedTimestamp || deployment.created_date > storedTimestamp) {
-    dispatch(setCurrentDeployment(deployment));
-    dispatch(setUpdateAvailable(true));
-  } else {
-    dispatch(setCurrentDeployment(deployment));
-    dispatch(setUpdateAvailable(false));
   }
 };
