+++ src/routers/rootRouter.ts
@@ -1,14 +1,16 @@
-import {handleDeploymentUpdate} from 'dispatchers/frontendDeployments';
+import {DEPLOYMENT_TIMESTAMP} from 'constants/localStorage';
 import {SocketDataType} from 'enums';
 import {getTradePriceChartData} from 'selectors/state';
 import {store} from 'store';
 import {setExchangeOrder} from 'store/exchangeOrders';
+import {setCurrentDeployment, setUpdateAvailable} from 'store/frontendDeployments';
 import {setNotification, setTotalUnreadCount} from 'store/notifications';
 import {updateOrderBookOrder} from 'store/orderBook';
 import {processTrade} from 'store/tradePriceChartData';
 import {setTrade} from 'store/trades';
 import {setWallet} from 'store/wallets';
 import {AppDispatch, RootState} from 'types';
+import {FrontendDeployment} from 'types/frontendDeployment';
 
 const handleCreateNotification = (dispatch: AppDispatch, socketData: any) => {
   dispatch(setNotification(socketData.notification));
@@ -31,6 +33,18 @@ const handleCreateTrade = (dispatch: AppDispatch, socketData: any) => {
   }
 };
 
+const handleDeploymentUpdate = (dispatch: AppDispatch, deployment: FrontendDeployment) => {
+  const storedTimestamp = localStorage.getItem(DEPLOYMENT_TIMESTAMP);
+
+  if (!storedTimestamp || deployment.created_date > storedTimestamp) {
+    dispatch(setCurrentDeployment(deployment));
+    dispatch(setUpdateAvailable(true));
+  } else {
+    dispatch(setCurrentDeployment(deployment));
+    dispatch(setUpdateAvailable(false));
+  }
+};
+
 const rootRouter = (dispatch: AppDispatch, event: MessageEvent) => {
   const socketData = JSON.parse(event.data);
   const {type} = socketData;
@@ -43,7 +57,7 @@ const rootRouter = (dispatch: AppDispatch, event: MessageEvent) => {
   } else if (type === SocketDataType.CREATE_TRADE) {
     handleCreateTrade(dispatch, socketData);
   } else if (type === SocketDataType.UPDATE_FRONTEND_DEPLOYMENT) {
-    dispatch(handleDeploymentUpdate(socketData.frontend_deployment));
+    handleDeploymentUpdate(dispatch, socketData.frontend_deployment);
   } else if (type === SocketDataType.UPDATE_WALLET) {
     dispatch(setWallet(socketData.wallet));
   }
